Add health bar ui

+++ Assets/Scripts/Weapons/Weapon.cs
@@ -3,7 +3,7 @@
 public abstract class Weapon : MonoBehaviour
 {
     protected Rigidbody2D weaponRb;
-    public int Damage = 1;
+    public float Damage = 1;
     public float CooldownPeriod = 0.1f;
 
     void Awake()
